staging: ethosu_arm: add Kconfig/Makefile and enable direct-drive mode
Introduce build infrastructure for the Arm Ethos-U NPU driver under
drivers/staging/ethosu_arm/, including:

- A new top-level Kconfig option ETHOSU_ARM (boolean) to enable the driver
  as built-in, with dependencies on OF and HAS_IOMEM.
- Optional sub-features: ETHOSU_DIRECT (enabled by default) for direct
  CPU-driven NPU control, and ETHOSU_RPMSG for RPMSG-based subsystem
  communication (disabled by default).
- Conditional compilation of RPMSG components based on CONFIG_ETHOSU_RPMSG.
- Updates to core driver files to wrap RPMSG-specific code in #ifdef
  guards.
- Minor build fixes: cast DMA_BIT_MASK to size_t and add 64-bit DMA
  address handling conditional on CONFIG_ARCH_DMA_ADDR_T_64BIT.

This change prepares the driver for user-mode memory mapping and direct
execution environments by making RPMSG optional and defaulting to the
lighter-weight direct-drive mode.

+++ drivers/staging/ethosu_arm/direct/ethosu_direct_device.c
@@ -389,7 +389,7 @@ static const struct ethosu_device_ops direct_ops = {
 
 static const struct ethosu_device_buffer_config buffer_config = {
 	.alignment = 16U,
-	.max_size  = DMA_BIT_MASK(40U),
+	.max_size  = (size_t)DMA_BIT_MASK(40U),
 };
 
 static int ethosu_set_npu_ops(struct platform_device *pdev)
